==> HW1/Code/2_Q2e.ipynb <==
from this_settings import *

## 1. Load data

data = pd.read_parquet(directory_cleandata+'data_a.parquet')
data_early = data[data['year']<=1989].copy()
data_late = data[data['year']>1989].copy()

data.head()

## 2. Full sample

reg_r = sm.OLS(endog=data['log_ret'],
               exog=sm.add_constant(data['log_pd_c_tm1']),
               missing='drop')\
        .fit()

reg_d = sm.OLS(endog=data['log_div_c_yoy'],
               exog=sm.add_constant(data['log_pd_c_tm1']),
               missing='drop')\
        .fit()

print(reg_r.summary())

print(reg_d.summary())

## 2. Early sample

reg_r = sm.OLS(endog=data_early['log_ret'],
               exog=sm.add_constant(data_early['log_pd_c_tm1']),
               missing='drop')\
        .fit()

reg_d = sm.OLS(endog=data_early['log_div_c_yoy'],
               exog=sm.add_constant(data_early['log_pd_c_tm1']),
               missing='drop')\
        .fit()

print(reg_r.summary())

print(reg_d.summary())

## 3. Late sample

reg_r = sm.OLS(endog=data_late['log_ret'],
               exog=sm.add_constant(data_late['log_pd_c_tm1']),
               missing='drop')\
        .fit()

reg_d = sm.OLS(endog=data_late['log_div_c_yoy'],
               exog=sm.add_constant(data_late['log_pd_c_tm1']),
               missing='drop')\
        .fit()

print(reg_r.summary())

print(reg_d.summary())
