==> HW1/Code/3_Q2f.ipynb <==
from this_settings import *

## 1. Load data

data = pd.read_parquet(directory_cleandata+'data_a.parquet')

## 2. Full sample

rho = data['pd_c'].mean() / (1.0+data['pd_c'].mean())

## Note: my rho is way off
rho

reg_r = sm.OLS(endog=data['log_ret'],
               exog=sm.add_constant(data['log_pd_c_tm1']),
               missing='drop')\
        .fit()

reg_d = sm.OLS(endog=data['log_div_c_yoy'],
               exog=sm.add_constant(data['log_pd_c_tm1']),
               missing='drop')\
        .fit()

reg_pd = sm.OLS(endog=data['log_pd_c'],
                exog=sm.add_constant(data['log_pd_c_tm1']),
                missing='drop')\
         .fit()

b_r = reg_r.params['log_pd_c_tm1']
b_d = reg_d.params['log_pd_c_tm1']
phi = reg_pd.params['log_pd_c_tm1']

phi

b_r_lr = b_r/(1.0-rho*phi)
b_d_lr = b_d/(1.0-rho*phi)

## Note, since Cochrane regresses on dividend yields, should be -1x this
b_r_lr

b_d_lr
