==> HW1/Code/4_Q3 copy.ipynb <==
from this_settings import *

## 1. Load data

## Import cleaned data
data = pd.read_parquet(directory_cleandata+'data_a.parquet')
data = data[(data['year']>=1946) &
            (data['year']<=2020)].copy()

## Demean the data to match Mankiw and Shapiro
## 
data.set_index('year',
               inplace=True)
data = data.copy()-data.mean()
data.reset_index(inplace=True)

## 2. Part a

## Regress returns on DP
reg_r = sm.OLS(endog=data['log_ret'],
               #exog=sm.add_constant(data['log_pd_c_tm1']),
               exog=data['log_pd_c_tm1'],
               missing='drop')\
        .fit()

## Regress PD on lagged PD
reg_pd = sm.OLS(endog=data['log_pd_c'],
                #exog=sm.add_constant(data['log_pd_c_tm1']),
                exog=data['log_pd_c_tm1'],
                missing='drop')\
         .fit()

## Grab coefficients
b_r_data = reg_r.params['log_pd_c_tm1']
phi_data = reg_pd.params['log_pd_c_tm1']

## Get coefficients
print(reg_r.summary())

## Get coefficients
print(reg_pd.summary())

## Get covariance matrix of residuals
e_data = pd.DataFrame({'e_r':reg_r.resid,
                       'e_pd':reg_pd.resid})
e_data_vcv = e_data.cov()

e_data_vcv

## 3. Part b

M=10000
N=75
p = e_data.corr().iloc[0,1]
e_sim_cov = np.array([[1,p],
                      [p,1]])

%%time
q3b = mankiw_shapiro(M=M,
                     N=N,
                     e_sim_cov=e_sim_cov,
                     phi_data=phi_data)

q3b['stats'].to_parquet('../Output/q3b_stats.parquet')
q3b['results'].to_parquet('../Output/q3b_results.parquet')

q3b['stats']

%%time
q3b_cov = mankiw_shapiro(M=M,
                     N=N,
                     e_sim_cov=e_data_vcv,
                     phi_data=phi_data)

## 4. Part c

e_sim_cov0 = np.array([[1,0],
                       [0,1]])

%%time
q3c = mankiw_shapiro(M=M,
                     N=N,
                     e_sim_cov=e_sim_cov0,
                     phi_data=phi_data)

print('here')

#q3c['stats'].to_parquet('../Output/q3c_stats.parquet')
#q3c['results'].to_parquet('../Output/q3c_results.parquet')

q3c_stats = pd.read_parquet('../Output/q3c_stats.parquet')

q3c_stats

##Don't forget to set the seed!
rng = np.random.default_rng(seed)

## Following Mankiw and Shapiro

## Draw e_r and e_pd from size-N bivarioate normal
## variances=1, covariance= corr(e_r,e_pd)
## ordered as e_r, e_pd
e_sim = rng.multivariate_normal(mean=[0,0],
                                cov=np.array([[1,p],
                                              [p,1]]),
                                size=N)
e_r_sim = e_sim[:,0]
e_pd_sim = e_sim[:,1]

## Set Y=v_t
## Set r = e_r_sim
r_sim = e_r_sim

## Generate X+t from eq 3 using innovations e
## Generate pd_t from pd AR(1) using innovations e_pd
## Initial value pd_0 is random from univariate normal, mean 0 and variance 1/(1-phi^2)
## Note: initial value gets used so we keep number of observations = N in the regression
## of pd(t) on pd(t-1)
pd_sim_0 = rng.normal(loc=0,
                      scale=np.sqrt(1.0/(1.0-phi_data**2))) # SD
pd_sim_panda = pd.DataFrame({'e_pd':e_pd_sim})
pd_sim_panda['pd'] = np.nan
pd_sim_panda.loc[0,'pd'] = pd_sim_0
for t in range(1,len(pd_sim_panda)+2):
    if t==0:
        pd_sim_tm1 = pd_sim_0
    if t>0:
        pd_sim_tm1 = pd_sim_panda.loc[t-1,'pd'].copy()
    if t==len(pd_sim_panda)+1:
        pd_sim_panda.loc[t,'pd'] = np.nan
        break
    pd_sim_panda.loc[t,'pd'] = phi_data * pd_sim_tm1 + pd_sim_panda.loc[t-1,'e_pd'].copy()
pd_sim_panda['pd_tm1'] = pd_sim_panda['pd'].copy().shift()
pd_sim_panda.dropna(subset=['pd','pd_tm1'],
                    inplace=True)
pd_sim = pd_sim_panda['pd'].copy().values
pd_sim_tm1 = pd_sim_panda['pd_tm1'].copy().values

## Estimate equation 2, grab t-stat
## Estimate return regression using generated data
reg_r_sim = sm.OLS(endog=r_sim,
                   exog=pd_sim,)\
            .fit()

## Estimate pd regression using generated data
reg_pd_sim = sm.OLS(endog=pd_sim,
                    exog=pd_sim_tm1,)\
             .fit()
             
## Grab regression coefficients
b_r_sim = reg_r_sim.params[0]
phi_sim = reg_pd_sim.params[0]

b_r_sim

b_r_sim

phi_sim

phi_sim

print(reg_r_sim.summary())

print(reg_pd_sim.summary())
